Address PR#95 comments.
* Split algebraic expansion/collection code out into own header.
* Split modcc expression -> algebraic representation code out from
  `test_kinetic_rewriter.cpp`
* Use rate symbol instead of rate expression in rewritten kinetic
  scheme ODE coefficients.
* Fix spelling errors.
* Fix crash bug (re-run semantic pass on cloned local assignment
  expression.)
